capture numeric httpstatuscode

+++ src/FilterLists.Services/SnapshotService/SnapshotDe.cs
@@ -58,7 +58,7 @@ namespace FilterLists.Services.SnapshotService
             }
             catch (WebException we)
             {
-                snapshot.HttpStatusCode = ((HttpWebResponse)we.Response).StatusCode.ToString();
+                snapshot.HttpStatusCode = ((int)((HttpWebResponse)we.Response).StatusCode).ToString();
                 return null;
             }
             catch (HttpRequestException)
@@ -73,7 +73,7 @@ namespace FilterLists.Services.SnapshotService
             using (var httpClient = new HttpClient())
             using (var httpResponseMessage = await httpClient.GetAsync(list.ViewUrl))
             {
-                snapshot.HttpStatusCode = httpResponseMessage.StatusCode.ToString();
+                snapshot.HttpStatusCode = ((int)httpResponseMessage.StatusCode).ToString();
                 if (httpResponseMessage.IsSuccessStatusCode)
                     return await httpResponseMessage.Content.ReadAsStringAsync();
             }
